train/Unsloth/requirements.txt
python-dotenv
wandb
unsloth
vllm
transformers==4.55.4